Publish new SDK version (#4734)
* Refactor useFleetAccessReview


* Export API constants


* Update provider flag to valid format


* Bump version to 0.3.5


* Fix useFleetAccessReview tests


---------
