+++ frontend/packages/multicluster-sdk/src/api/constants.ts
@@ -1,2 +1,2 @@
 /* Copyright Contributors to the Open Cluster Management project */
-export const REQUIRED_PROVIDER_FLAG = '@stolostron/multicluster-sdk/provider@1'
+export const REQUIRED_PROVIDER_FLAG = 'MULTICLUSTER_SDK_PROVIDER_1'
