--- frontend/packages/multicluster-sdk/src/api/useFleetAccessReview/index.ts
@@ -1,2 +1,0 @@
-/* Copyright Contributors to the Open Cluster Management project */
-export { useFleetAccessReview } from './access-review'
