+++ frontend/src/utils/flags/useFeatureFlags.ts
@@ -1,6 +1,6 @@
 /* Copyright Contributors to the Open Cluster Management project */
 import { SetFeatureFlag } from '@openshift-console/dynamic-plugin-sdk'
-
+import { REQUIRED_PROVIDER_FLAG } from '@stolostron/multicluster-sdk'
 import { MultiClusterHubComponent } from '../../resources/multi-cluster-hub-component'
 import { getBackendUrl, getRequest } from '../../resources/utils'
 import { FEATURE_FLAGS } from './consts'
@@ -16,7 +16,7 @@ const useFeatureFlags = (setFeatureFlag: SetFeatureFlag) => {
     )
   )
   // Supports first version of multicluster SDK provider requirements
-  setFeatureFlag('@stolostron/multicluster-sdk/provider@1', true)
+  setFeatureFlag(REQUIRED_PROVIDER_FLAG, true)
 }
 
 export default useFeatureFlags
